requirements.txt
streamlit
pandas
pymongo[srv]
gcsfs
plotly
apache-airflow-providers-google